PRMD-Code-master/requirements.txt
azure-iot-device==2.14.0
certifi==2025.4.26
charset-normalizer==3.4.2
deprecation==2.1.0
dotenv==0.9.9
hx711==1.1.2.3
idna==3.10
janus==2.0.0
lgpio==0.2.2.0
packaging==25.0
paho-mqtt==1.6.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-unixsocket2==1.0.0
smbus2==0.5.0
typing_extensions==4.13.2
urllib3==2.4.0
